move to bot package and clean up kernel conf
